+++ Firmware/Master/src/network/udpclient.h
@@ -41,7 +41,7 @@ namespace ServerConnection
     int DataReady();
     bool IsHapticServerConnected();
     bool IsSlimeServerConnected();
-    uint32_t getUnixTime();
+    unsigned long getUnixTime();
 
 }
